Make FailedTestResultListener optional
